fix(landing): move git_topology primitive to kernel (fixes charter layer inversion, #3373)
WP01 unified the git-common-dir/toplevel probe into
src/specify_cli/git/git_topology.py and repointed charter.resolution to
consume it — but charter is a LOWER layer than specify_cli
(kernel <- doctrine <- charter <- specify_cli), so the import inverted the
layering (test_layer_rules::TestCharterBoundary,
test_charter_no_specify_cli_import: "src.charter.resolution imports
specify_cli.git.git_topology").

The primitive is pure-stdlib (subprocess/functools/pathlib — zero landscape
deps), so it belongs in kernel, "the zero-dependency root" that already homes
shared git plumbing (kernel/vcs_lock.py). Move it there; all four consumers
(charter.resolution + the three specify_cli sites) import from kernel.git_topology,
which every layer may legally reach. WP01's single-copy unification is preserved
(one primitive, one canonicalization contract).

Update the one-copy gate's path pin and the test patch targets to the kernel
home. Layer gates, one-copy gate, git_topology parity, and charter canonical-root
resolution all green.
