+++ src/kernel/git_topology.py
